Fix chain-resource matching and DONE transition
The algorithm to match chains was failing in common scenarios, so it has
been reworked to do a single query matching the resources passed in.
From the chains found, find the ones with all matching resources and
return the result. It shouldn't find more than one, but if it does,
it'll throw an error.

Also, when a DONE state is found, we need to transition to that DONE
state. DONE is a special case with an early exit (no callback).