Add option to resources to query a subset
Depending on the circumstance, sometimes we don't want to match all the
resources, just as many as are provided.

For example, say we don't know all the resources involved in an
evaluation:

```
rsc_map = [
    ('profiles', 'abc123'),
    ('departure_services', '123abc')
]
```

But the following resources exist:

```
chain.resources = [
    {
        'resource_type': 'profiles',
        'resource_id': 'abc123'
    },
    {
        'resource_type': 'departure_services',
        'resource_id': '123abc'
    },
    {
        'resource_type': 'evaluations',
        'resource_id': 'evalID'
    },
]
```

If we were to query the rsc_map with `can_be_subset = False`,
the chain would not be found because it doesn't match the resources
exactly. With `can_be_subset = True`, it'll find that chain.resources
contains all of the rsc_map resources and will mark it as found.